+++ contracts/test/consensus/Consensus-ValidatorRegistration.sol
@@ -35,7 +35,7 @@ contract ConsensusTest is Base {
         assertEq(consensus.activeValidatorsCount(), 0);
 
         // Set a custom fee for validator registration
-        uint256 customFee = 40 ether;
+        uint128 customFee = 40 ether;
         consensus.setFee(customFee);
 
         address payable addr = payable(address(1));
